+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/update/SimpleUpdateContext.java
@@ -61,6 +61,11 @@ public class SimpleUpdateContext implements UpdateContext {
     }
 
     @Override
+    public boolean isForceEntity() {
+        return false;
+    }
+
+    @Override
     public boolean addVersionCheck(Class<?> entityClass, Object id) {
         return false;
     }